refactor: move PackageNamespaceAndShortName to questionpy_common and add from_string
